Styler updates, added get_stocking_events and updated readme
The commit includes small updates to most files that were
automatically performed by style_pkg().  It also includes the function
get_stocking_events() which was missed in earlier commits.  Finally,
README.rmd was updated and now includes actual R-output in all
sections. Is version of the package is not yet ready for production -
the get_root_api() function currently expects the api to be available
on local host.