Added get_stocking_events, max page count, development env
This commit actually adds get_stocking_events that was somehow missed
on the last commit (again), added an if statement to get_root_url so
that a local versions of the api can be used in development, and a
limit on the number of iterative pages can be requested by
api_2_dataframe. If the number of pages exceeds the limit, a warning
is issued and provides some suggestions on how to fetch more data if
necessary.