Added functions to show api filters
This commit adds several functions that are used to check and show
available api filters. Any functions associated with a filterable api
endpoint were updated to check the supplied filter and report any
issues, new functions were added to display the available filters, and
the read-me file was updated and regenerated.